nodes/haptic_controller/pyproject.toml
[tool.poetry]
name = "haptic-controller"
version = "0.1.0"
description = "Haptic controller: force-feedback resistance and zero-G hold mode for leader gripper (gripper-only pilot)"
readme = "README.md"
packages = [{ include = "haptic_controller", from = "." }]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = ">=6.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^24.0.0"
isort = "^5.13.0"
autoflake = "^2.3.1"
vulture = "^2.11"
poethepoet = "^0.42.0"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[tool.flake8]
max-line-length = 120

[tool.vulture]
exclude = ["**/__pycache__"]
min_confidence = 80

[tool.poe.tasks]
lint = { shell = """
  set -e
  flake8 haptic_controller
  black --check haptic_controller tests
  isort --check haptic_controller tests
  autoflake --check --recursive --remove-all-unused-imports --expand-star-imports --remove-unused-variables haptic_controller tests
  vulture haptic_controller tests --min-confidence 80
""" }
lint-fix = { shell = """
  black haptic_controller tests
  isort haptic_controller tests
  autoflake --in-place --recursive --remove-all-unused-imports --expand-star-imports --remove-unused-variables haptic_controller tests
""" }
test = { shell = "pytest tests/ -v" }

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
